% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pythagorean.R
\name{Pythagorean}
\alias{Pythagorean}
\title{The Pythagorean Theorem}
\usage{
Pythagorean(Side1, Side2, Hyp)
}
\arguments{
\item{Side1}{is the length of one side.}

\item{Side2}{is the length of another side.}

\item{Hyp}{indicates whether the hypotenuse is included in one of the two provided sides.
The computation changes whether this parameter is TRUE or FALSE.}
}
\description{
This function allows you to compute any side of a right triangle using a famous theorem.
}
\examples{
Pythagorean(3,4,T)
Pythagorean(3,4,F)
}
\keyword{Triangles}
