% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimmed_Mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{The Trimmed Mean}
\usage{
trimmed_mean(x, s, l)
}
\arguments{
\item{s}{is the number of low values to cut off.}

\item{l}{is the number of high values to cut off.}

\item{X}{is the vector of numeric values.}
}
\description{
This function allows you to compute the trimmed mean of a numeric vector.
}
\examples{
Trimmed_Mean(iris$Sepal.Length, 1, 1)
}
\keyword{Mean}
\keyword{Trimmed}
